fix contact detail drawer
